@@ -1,6 +1,6 @@
 {
   "name": "@hoprnet/hopr-ethereum",
-  "version": "0.4.0-testnet",
+  "version": "0.4.0",
   "description": "On-chain logic for hopr.network",
   "homepage": "https://hopr.network",
   "repository": {
@@ -10,7 +10,7 @@
   "license": "LGPL-3.0-only",
   "main": "./build/lib/scripts/api.js",
   "files": [
-    "truffle-networks.json",
+    "truffle-networks.js",
     "truffle-config.js",
     "contracts/",
     "build/extracted/",
@@ -18,43 +18,46 @@
     "migrations/"
   ],
   "engines": {
-    "yarn": "^1.19.2",
-    "node": "^12"
+    "yarn": "1.19.2",
+    "node": "12.9.1"
   },
   "scripts": {
     "build": "ts-node -T ./scripts/cli.ts build",
     "coverage": "ts-node -T ./scripts/cli.ts coverage",
     "fund": "ts-node -T ./scripts/cli.ts fund",
     "migrate": "ts-node -T ./scripts/cli.ts migrate",
+    "verify": "ts-node -T ./scripts/cli.ts verify",
     "network": "ts-node -T ./scripts/cli.ts network",
     "test": "ts-node -T ./scripts/cli.ts test",
-    "format": "prettier --check .",
-    "prepare": "yarn build"
+    "lint": "prettier --check .",
+    "prepublish": "yarn build"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "yarn format"
+      "pre-commit": "yarn lint"
     }
   },
   "dependencies": {
-    "@hoprnet/hopr-demo-seeds": "1.0.1",
-    "@hoprnet/hopr-utils": "0.1.3",
+    "@hoprnet/hopr-demo-seeds": "1.0.3-refactor.3ad82ff",
+    "@hoprnet/hopr-utils": "0.1.6-refactor.61df948",
     "@openzeppelin/contracts": "^3.0.1",
     "@openzeppelin/test-helpers": "^0.5.5",
     "@truffle/debug-utils": "^4.1.1",
     "@truffle/hdwallet-provider": "^1.0.34",
     "bignumber.js": "^9.0.0",
-    "ganache-core": "^2.10.2",
+    "dotenv": "^8.2.0",
     "truffle": "^5.1.23",
     "truffle-plugin-verify": "^0.3.10",
     "ts-node": "^8.9.1",
     "web3": "^1.2.7"
   },
   "devDependencies": {
+    "@hoprnet/hopr-testing": "0.1.2",
     "@types/chai": "^4.2.11",
     "@types/mocha": "^7.0.2",
     "@types/node": "^13.13.4",
     "chai": "^4.2.0",
+    "ganache-core": "^2.10.2",
     "husky": "^4.2.5",
     "mocha": "^7.1.2",
     "prettier": "^2.0.5",
