@@ -1,12 +1,13 @@
 import Web3 from 'web3'
 import HoprTokenAbi from '../../build/extracted/abis/HoprToken.json'
-import networks from '../../truffle-networks.json'
+import { HOPR_TOKEN } from '../addresses'
+import networks from '../../truffle-networks'
 
 const AMOUNT = Web3.utils.toWei('1000000', 'ether')
 
 export default async (amount?: number) => {
   const web3 = new Web3(`ws://${networks.development.host}:${networks.development.port}`)
-  const hoprToken: any = new web3.eth.Contract(HoprTokenAbi as any, '0x66DB78F4ADD912a6Cb92b672Dfa09028ecc3085E')
+  const hoprToken: any = new web3.eth.Contract(HoprTokenAbi as any, HOPR_TOKEN.private)
 
   const accounts = await web3.eth.getAccounts()
   const owner = accounts[0]
