@@ -1,10 +1,15 @@
 import verify from './verify'
-import { bash, isLocalNetwork, getContractNames } from './utils'
+import { bash } from './utils'
+import networks from '../../truffle-networks'
 
 export default async (network: string = 'development') => {
   await bash(`npx truffle migrate --network ${network}`)
 
-  if (!isLocalNetwork(network)) {
-    await verify(network, ...getContractNames())
+  const config = networks[network]
+
+  if (config.network_type === 'mainnet') {
+    await verify(network, 'HoprToken', 'HoprChannels', 'HoprMinter')
+  } else if (config.network_type === 'testnet') {
+    await verify(network, 'HoprToken', 'HoprChannels', 'HoprFaucet')
   }
 }
