@@ -1,7 +1,6 @@
-import ganacheCore from './ganache-core'
 import truffleTypings from './truffle-typings'
 import typechain from './typechain'
 
 export default () => {
-  return Promise.all([ganacheCore(), truffleTypings(), typechain()])
+  return Promise.all([truffleTypings(), typechain()])
 }
